Add HTML dashboard rendering utility for resource dashboard
- Introduced `dashboard_html_main.py` containing the `build_html_dashboard` function to generate a self-contained HTML file for visualizing test run data with Plotly.js charts and CPU suggestions.
- Updated `tests_resource_dashboard.py` to utilize the new HTML dashboard function, enhancing the resource dashboard's capabilities.
- Removed the previous HTML dashboard implementation from `tests_resource_dashboard.py` to streamline the codebase.
